+++ Source/PlaySequencer.h
@@ -66,7 +66,7 @@ public:
 
    //IGridControllerListener
    void OnControllerPageSelected() override;
-   void OnGridButton(int x, int y, float velocity, IGridController* grid) override;
+   void OnGridButton(int x, int y, double velocity, IGridController* grid) override;
 
    //ITimeListener
    void OnTimeEvent(double time) override;
