+++ Source/SamplerGrid.h
@@ -65,7 +65,7 @@ public:
 
    //IGridControllerListener
    void OnControllerPageSelected() override;
-   void OnGridButton(int x, int y, float velocity, IGridController* grid) override;
+   void OnGridButton(int x, int y, double velocity, IGridController* grid) override;
 
    //INoteReceiver
    void PlayNote(NoteMessage note) override;
