+++ Source/StutterControl.cpp
@@ -260,7 +260,7 @@ void StutterControl::UpdateGridLights()
    }
 }
 
-bool StutterControl::OnPush2Control(Push2Control* push2, MidiMessageType type, int controlIndex, float midiValue)
+bool StutterControl::OnPush2Control(Push2Control* push2, MidiMessageType type, int controlIndex, double midiValue)
 {
    if (type == kMidiMessage_Note)
    {
